requirements.txt
kafka-python==2.0.2
six==1.16.0
psycopg2-binary==2.9.1  # or any other dependencies you're using
